Property integers are 64-bit end to end
PropNode stores int64, the PData descriptor and the C API
(add_property, has_property, get_property_value) carry int64, the
store writes and reads 64-bit JSON integers, the flat importer
parses with strtoll, and every display or serialization site prints
long long. Existing 32-bit values round-trip byte-identically. MUF
getpropval still narrows to the interpreter's 32-bit int until the
64-bit MUF integer conversion lands, at which point full values flow
through unchanged; the offline harness proves a 64-bit value stores
and reads back exactly. MUF suite 135 of 135, round trip identical.

+++ src/ObjectStore.cpp
@@ -339,7 +339,7 @@ propsFromJson(dbref obj, const json &arr)
                 pdat.data.lok = parse_boolexp(-1, (dbref) 1, sval.c_str(), 32767);
                 break;
             case PROP_INTTYP:
-                pdat.data.val = v.get<int>();
+                pdat.data.val = v.get<int64_t>();
                 break;
             case PROP_FLTTYP:
                 pdat.data.fval = v.get<double>();
